refactor(src/test/blackjack/domain/): 패키지 추가
실제 구현 코드와 패키지 구조를 통일 시키기 위해 role, card 패키지 추가

BREAKING CHANGE: src/test/blackjack/domain 하위에 role, card 패키지 추가
