+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
@@ -9,9 +9,6 @@ import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Deck;
 
 @DisplayName("Deck 태스트")
 class DeckTest {
