+++ src/test/java/blackjack/domain/card/HandTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
@@ -11,8 +11,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Hand;
 import blackjack.domain.factory.CardMockFactory;
 import blackjack.domain.util.CreateHand;
